refactor(client): extract UI state management from App component into useJobSearch hook
